Merged

+++ app/models/spree/order_decorator.rb
@@ -80,6 +80,8 @@ module Spree
     # ensure that user has sufficient credits to cover adjustments
     #
     def ensure_sufficient_credit
+      # update only if this order is still incomplete or transitioning confirm->complete.
+      return unless !completed? or changed_attributes['state'] == 'confirm'
       if user.store_credits_total < store_credit_amount
         # user's credit does not cover all adjustments.
         adjustments.store_credits.destroy_all
